Allow user to configure import export parameters for loadflow iTools
